--- lib/segment/src/index/field_index/map_index/mmap_map_index.rs
@@ -1,497 +1,0 @@
-use std::borrow::{Borrow, Cow};
-use std::iter;
-use std::ops::BitOrAssign;
-use std::path::{Path, PathBuf};
-
-use ahash::HashMap;
-use common::bitvec::{BitSlice, BitSliceExt, BitVec};
-use common::counter::conditioned_counter::ConditionedCounter;
-use common::counter::hardware_counter::HardwareCounterCell;
-use common::counter::iterator_hw_measurement::HwMeasurementIteratorExt;
-use common::fs::{atomic_save_json, clear_disk_cache, read_json};
-use common::mmap::create_and_ensure_length;
-use common::persisted_hashmap::{Key, READ_ENTRY_OVERHEAD, UniversalHashMap, serialize_hashmap};
-use common::stored_bitslice::MmapBitSlice;
-use common::types::PointOffsetType;
-use common::universal_io::{MmapFile, OpenOptions};
-use fs_err as fs;
-use itertools::Itertools;
-use serde::{Deserialize, Serialize};
-
-use super::{IdIter, MapIndexKey};
-use crate::common::Flusher;
-use crate::common::operation_error::{OperationError, OperationResult};
-use crate::index::field_index::stored_point_to_values::{StoredPointToValues, ValuesIter};
-
-const DELETED_PATH: &str = "deleted.bin";
-const HASHMAP_PATH: &str = "values_to_points.bin";
-const CONFIG_PATH: &str = "mmap_field_index_config.json";
-
-/// Mmap-backed immutable map index.
-///
-/// On-disk state (`values_to_points.bin`, `deleted.bin`, `point_to_values.*`,
-/// `mmap_field_index_config.json`) is written once during [`Self::build`] and
-/// not mutated afterwards: `deleted.bin` records only the points whose payload
-/// was empty at build time.
-///
-/// Runtime deletions live in the in-memory `Storage::deleted` bitvec. They are
-/// **not persisted** — [`Self::flusher`] is a no-op and [`Self::remove_point`]
-/// only updates the in-memory bitvec. Callers must re-supply the authoritative
-/// deletion set (typically `id_tracker.deleted_point_bitslice()`) via the
-/// `deleted_points` argument to [`Self::open`] on reload.
-pub struct MmapMapIndex<N: MapIndexKey + Key + ?Sized> {
-    path: PathBuf,
-    pub(super) storage: Storage<N>,
-    deleted_count: usize,
-    total_key_value_pairs: usize,
-    is_on_disk: bool,
-}
-
-pub(super) struct Storage<N: MapIndexKey + Key + ?Sized> {
-    pub(super) value_to_points: UniversalHashMap<N, PointOffsetType, MmapFile>,
-    point_to_values: StoredPointToValues<N, MmapFile>,
-    /// In-memory deletion bitmap. Reconstructed at load time as the union of
-    /// the build-time empty-payload bits read from `deleted.bin` and the
-    /// segment-level deleted bitslice supplied by the id-tracker. Not persisted.
-    pub(super) deleted: BitVec,
-}
-
-impl<N: MapIndexKey + Key + ?Sized> Storage<N> {
-    pub(crate) fn ram_usage_bytes(&self) -> usize {
-        let Self {
-            value_to_points: _,
-            point_to_values,
-            deleted,
-        } = self;
-
-        // `value_to_points` is a mmap-backed hashmap with no in-memory state.
-        point_to_values.ram_usage_bytes() + deleted.capacity().div_ceil(u8::BITS as usize)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct MmapMapIndexConfig {
-    total_key_value_pairs: usize,
-}
-
-impl<N: MapIndexKey + Key + ?Sized> MmapMapIndex<N> {
-    /// Open and load mmap map index from the given path
-    pub fn open(
-        path: &Path,
-        is_on_disk: bool,
-        deleted_points: &BitSlice,
-    ) -> OperationResult<Option<Self>> {
-        let hashmap_path = path.join(HASHMAP_PATH);
-        let deleted_path = path.join(DELETED_PATH);
-        let config_path = path.join(CONFIG_PATH);
-
-        // If config doesn't exist, assume the index doesn't exist on disk
-        if !config_path.is_file() {
-            return Ok(None);
-        }
-
-        let config: MmapMapIndexConfig = read_json(&config_path)?;
-
-        let do_populate = !is_on_disk;
-
-        let value_to_points = UniversalHashMap::open(
-            &hashmap_path,
-            OpenOptions {
-                writeable: false,
-                populate: Some(do_populate),
-                ..OpenOptions::default()
-            },
-        )?;
-        let point_to_values = StoredPointToValues::open(path, do_populate)?;
-
-        let mut deleted = deleted_points.to_owned();
-
-        let deleted_payload_mmap = MmapBitSlice::open(&deleted_path, OpenOptions::default())?;
-        let deleted_payloads_bitslice = deleted_payload_mmap.read_all()?;
-
-        // `deleted` length must match `point_to_values.len()` because it only
-        // tracks the index's contents. The id-tracker's deleted mask can be
-        // shorter or longer; if shorter, the missing entries default to live
-        // (the id-tracker is the source of truth for deletions, and a shorter
-        // mask just means it doesn't yet know about those higher offsets).
-        deleted.resize(point_to_values.len(), false);
-        deleted.bitor_assign(deleted_payloads_bitslice.as_ref());
-
-        let deleted_count = deleted.count_ones();
-
-        Ok(Some(Self {
-            path: path.to_path_buf(),
-            storage: Storage {
-                value_to_points,
-                point_to_values,
-                deleted,
-            },
-            deleted_count,
-            total_key_value_pairs: config.total_key_value_pairs,
-            is_on_disk,
-        }))
-    }
-
-    pub fn build(
-        path: &Path,
-        point_to_values: Vec<Vec<<N as MapIndexKey>::Owned>>,
-        values_to_points: HashMap<<N as MapIndexKey>::Owned, Vec<PointOffsetType>>,
-        is_on_disk: bool,
-        deleted_points: &BitSlice,
-    ) -> OperationResult<Self> {
-        fs::create_dir_all(path)?;
-
-        let hashmap_path = path.join(HASHMAP_PATH);
-        let deleted_path = path.join(DELETED_PATH);
-        let config_path = path.join(CONFIG_PATH);
-
-        atomic_save_json(
-            &config_path,
-            &MmapMapIndexConfig {
-                total_key_value_pairs: point_to_values.iter().map(|v| v.len()).sum(),
-            },
-        )?;
-
-        serialize_hashmap(
-            &hashmap_path,
-            values_to_points
-                .iter()
-                .map(|(value, ids)| (value.borrow(), ids.iter().copied())),
-        )?;
-
-        StoredPointToValues::<N, MmapFile>::from_iter(
-            path,
-            point_to_values.iter().enumerate().map(|(idx, values)| {
-                (
-                    idx as PointOffsetType,
-                    values.iter().map(|value| value.borrow()),
-                )
-            }),
-        )?;
-
-        {
-            let deleted_flags_count = point_to_values.len();
-            let _ = create_and_ensure_length(
-                &deleted_path,
-                deleted_flags_count
-                    .div_ceil(u8::BITS as usize)
-                    .next_multiple_of(size_of::<u64>()),
-            )?;
-
-            let mut deleted = MmapBitSlice::open(&deleted_path, OpenOptions::default())?;
-            deleted.set_ascending_bits_batch(
-                point_to_values
-                    .iter()
-                    .enumerate()
-                    .filter(|(_, values)| values.is_empty())
-                    .map(|(idx, _)| (idx as u64, true)),
-            )?;
-            deleted.flusher()()?;
-        }
-
-        Self::open(path, is_on_disk, deleted_points)?.ok_or_else(|| {
-            OperationError::service_error("Failed to open MmapMapIndex after building it")
-        })
-    }
-
-    /// No-op flusher: the on-disk state is build-time only. See the type-level
-    /// docs on [`MmapMapIndex`] for the deletion durability contract.
-    pub fn flusher(&self) -> Flusher {
-        Box::new(|| Ok(()))
-    }
-
-    pub fn wipe(self) -> OperationResult<()> {
-        let files = self.files();
-        let path = self.path.clone();
-        // drop mmap handles before deleting files
-        drop(self);
-        for file in files {
-            fs::remove_file(file)?;
-        }
-        let _ = fs::remove_dir(path);
-        Ok(())
-    }
-
-    pub fn files(&self) -> Vec<PathBuf> {
-        let mut files = vec![
-            self.path.join(HASHMAP_PATH),
-            self.path.join(DELETED_PATH),
-            self.path.join(CONFIG_PATH),
-        ];
-        files.extend(self.storage.point_to_values.files());
-        files
-    }
-
-    pub fn immutable_files(&self) -> Vec<PathBuf> {
-        let mut files = vec![
-            self.path.join(HASHMAP_PATH),
-            self.path.join(DELETED_PATH),
-            self.path.join(CONFIG_PATH),
-        ];
-        files.extend(self.storage.point_to_values.immutable_files());
-        files
-    }
-
-    /// Marks `idx` as deleted in the in-memory deletion bitvec.
-    ///
-    /// Not persisted: on reopen, deletions must be re-supplied via the
-    /// `deleted_points` argument to [`Self::open`].
-    pub fn remove_point(&mut self, idx: PointOffsetType) {
-        let idx = idx as usize;
-        if idx < self.storage.deleted.len() && !self.storage.deleted.get_bit(idx).unwrap_or(true) {
-            self.storage.deleted.set(idx, true);
-            self.deleted_count += 1;
-        }
-    }
-
-    pub fn check_values_any(
-        &self,
-        idx: PointOffsetType,
-        hw_counter: &HardwareCounterCell,
-        check_fn: impl Fn(&N) -> bool,
-    ) -> OperationResult<bool> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-
-        // Measure self.deleted access.
-        hw_counter
-            .payload_index_io_read_counter()
-            .incr_delta(size_of::<bool>());
-
-        let is_deleted = self
-            .storage
-            .deleted
-            .get_bit(idx as usize)
-            .is_some_and(|b| b);
-
-        Ok(!is_deleted
-            && self
-                .storage
-                .point_to_values
-                .check_values_any(idx, |v| check_fn(v), &hw_counter)?)
-    }
-
-    pub fn get_values(
-        &self,
-        idx: PointOffsetType,
-        hw_counter: &HardwareCounterCell,
-    ) -> Option<Box<dyn Iterator<Item = Cow<'_, N>> + '_>> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-
-        // We can account cost of reading `bool`, but it will likely be more expensive, than
-        // actually reading bool itself.
-
-        if self.storage.deleted.get_bit(idx as usize) == Some(false) {
-            self.storage
-                .point_to_values
-                .values_iter(idx, hw_counter)
-                .ok()?
-                .map(|iter| Box::new(iter) as Box<dyn Iterator<Item = Cow<'_, N>>>)
-        } else {
-            None
-        }
-    }
-
-    pub fn for_points_values(
-        &self,
-        mut points: impl Iterator<Item = PointOffsetType>,
-        hw_counter: &HardwareCounterCell,
-        mut f: impl FnMut(PointOffsetType, ValuesIter<'_, N>),
-    ) -> OperationResult<()> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-
-        points.try_for_each(|idx| {
-            if self.storage.deleted.get_bit(idx as usize) != Some(false) {
-                return Ok(());
-            }
-            if let Some(iter) = self.storage.point_to_values.values_iter(idx, hw_counter)? {
-                f(idx, iter);
-            }
-            Ok(())
-        })
-    }
-
-    pub fn values_count(&self, idx: PointOffsetType) -> Option<usize> {
-        if self.storage.deleted.get_bit(idx as usize) == Some(false) {
-            self.storage.point_to_values.get_values_count(idx).ok()?
-        } else {
-            None
-        }
-    }
-
-    pub fn get_indexed_points(&self) -> usize {
-        self.storage
-            .point_to_values
-            .len()
-            .saturating_sub(self.deleted_count)
-    }
-
-    /// Returns the number of key-value pairs in the index.
-    /// Note that is doesn't count deleted pairs.
-    pub fn get_values_count(&self) -> usize {
-        self.total_key_value_pairs
-    }
-
-    pub fn get_unique_values_count(&self) -> usize {
-        self.storage.value_to_points.keys_count()
-    }
-
-    pub fn get_count_for_value(
-        &self,
-        value: &N,
-        hw_counter: &HardwareCounterCell,
-    ) -> Option<usize> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-
-        // Since `value_to_points.get` doesn't actually force read from disk for all values
-        // we need to only account for the overhead of hashmap lookup
-        hw_counter
-            .payload_index_io_read_counter()
-            .incr_delta(READ_ENTRY_OVERHEAD);
-
-        match self
-            .storage
-            .value_to_points
-            .unbatched_get_values_count(value)
-        {
-            Ok(Some(count)) => Some(count),
-            Ok(None) => None,
-            Err(err) => {
-                debug_assert!(
-                    false,
-                    "Error while getting count for value {value:?}: {err:?}",
-                );
-                log::error!("Error while getting count for value {value:?}: {err:?}");
-                None
-            }
-        }
-    }
-
-    pub fn get_iterator(&self, value: &N, hw_counter: &HardwareCounterCell) -> IdIter<'_> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-
-        match self.storage.value_to_points.unbatched_get(value) {
-            Ok(Some(values)) => {
-                // We're iterating over the whole (mmapped) slice
-                hw_counter
-                    .payload_index_io_read_counter()
-                    .incr_delta(size_of_val(values.as_slice()) + READ_ENTRY_OVERHEAD);
-
-                Box::new(
-                    values.into_iter().filter(|idx| {
-                        !self.storage.deleted.get_bit(*idx as usize).unwrap_or(false)
-                    }),
-                )
-            }
-            Ok(None) => {
-                hw_counter
-                    .payload_index_io_read_counter()
-                    .incr_delta(READ_ENTRY_OVERHEAD);
-
-                Box::new(iter::empty())
-            }
-            Err(err) => {
-                debug_assert!(
-                    false,
-                    "Error while getting iterator for value {value:?}: {err:?}",
-                );
-                log::error!("Error while getting iterator for value {value:?}: {err:?}");
-                Box::new(iter::empty())
-            }
-        }
-    }
-
-    pub fn for_each_value(&self, f: impl FnMut(&N) -> OperationResult<()>) -> OperationResult<()> {
-        self.storage.value_to_points.for_each_key(f)
-    }
-
-    pub fn for_each_count_per_value(
-        &self,
-        deferred_internal_id: Option<PointOffsetType>,
-        mut f: impl FnMut(&N, usize) -> OperationResult<()>,
-    ) -> OperationResult<()> {
-        self.storage.value_to_points.for_each_entry(|k, v| {
-            let count = v
-                .iter()
-                .filter(|&&idx| {
-                    !self.storage.deleted.get_bit(idx as usize).unwrap_or(true)
-
-                    // TODO(deferred): Maybe we can improve this filter and use take_while instead. For this we
-                    // need to make sure that `v` is always sorted which we _can_ enforce when finalizing the index.
-                    && deferred_internal_id.is_none_or(|deferred| idx < deferred)
-                })
-                .unique()
-                .count();
-            f(k, count)
-        })
-    }
-
-    pub fn for_each_value_map(
-        &self,
-        hw_counter: &HardwareCounterCell,
-        mut f: impl FnMut(&N, &mut dyn Iterator<Item = PointOffsetType>) -> OperationResult<()>,
-    ) -> OperationResult<()> {
-        let hw_counter = self.make_conditioned_counter(hw_counter);
-        let deleted = &self.storage.deleted;
-
-        self.storage.value_to_points.for_each_entry(|k, v| {
-            hw_counter
-                .payload_index_io_read_counter()
-                .incr_delta(k.write_bytes());
-
-            let mut iter = v
-                .iter()
-                .copied()
-                .filter(|idx| !deleted.get_bit(*idx as usize).unwrap_or(true))
-                .measure_hw_with_acc(
-                    hw_counter.new_accumulator(),
-                    size_of::<PointOffsetType>(),
-                    |i| i.payload_index_io_read_counter(),
-                );
-
-            f(k, &mut iter)
-        })
-    }
-
-    fn make_conditioned_counter<'a>(
-        &self,
-        hw_counter: &'a HardwareCounterCell,
-    ) -> ConditionedCounter<'a> {
-        ConditionedCounter::new(self.is_on_disk, hw_counter)
-    }
-
-    pub fn is_on_disk(&self) -> bool {
-        self.is_on_disk
-    }
-
-    /// Populate all pages in the mmap.
-    /// Block until all pages are populated.
-    pub fn populate(&self) -> OperationResult<()> {
-        self.storage.value_to_points.populate()?;
-        self.storage.point_to_values.populate()?;
-        Ok(())
-    }
-
-    /// Drop disk cache.
-    pub fn clear_cache(&self) -> OperationResult<()> {
-        let Self {
-            path,
-            storage,
-            deleted_count: _,
-            total_key_value_pairs: _,
-            is_on_disk: _,
-        } = self;
-        let Storage {
-            value_to_points,
-            point_to_values,
-            deleted: _,
-        } = storage;
-        value_to_points.clear_ram_cache()?;
-        clear_disk_cache(&path.join(DELETED_PATH))?;
-        point_to_values.clear_cache()?;
-        Ok(())
-    }
-
-    pub(crate) fn ram_usage_bytes(&self) -> usize {
-        self.storage.ram_usage_bytes()
-    }
-}
